Split transaction cost source context mapping
