+++ src/api/services/construction_source_product_context.py
@@ -13,11 +13,12 @@ from src.api.services.construction_liquidity_source_context import (
     source_liquidity_context,
     source_status_to_method_status,
 )
+from src.api.services.construction_transaction_cost_source_context import (
+    transaction_cost_context_from_curve,
+)
 from src.core.construction.models import (
     AuthoritativeCurrencyOverlayContext,
     AuthoritativeExecutionAcknowledgementContext,
-    AuthoritativeTransactionCostContext,
-    AuthoritativeTransactionCostPoint,
     ConstructionAuthorityContext,
 )
 from src.core.dpm_source_context import (
@@ -28,52 +29,8 @@ from src.core.dpm_source_context import (
     DpmCoreExternalHedgePolicyResponse,
     DpmCoreExternalOrderExecutionAcknowledgementResponse,
     DpmCoreExecutionContext,
-    DpmCoreTransactionCostCurveResponse,
     DpmResolvedSourceContext,
 )
-
-
-def transaction_cost_context_from_curve(
-    curve: DpmCoreTransactionCostCurveResponse,
-) -> AuthoritativeTransactionCostContext:
-    curve_payload = curve.model_dump(mode="json", exclude_none=True)
-    source_hash = hash_canonical_payload(curve_payload)
-    source_id = (
-        curve.source_batch_fingerprint
-        or curve.lineage.get("source_batch_fingerprint")
-        or curve.page.request_scope_fingerprint
-    )
-    return AuthoritativeTransactionCostContext(
-        supportability_status=source_status_to_method_status(curve.supportability.state),
-        source_system="lotus-core",
-        source_product_name=curve.product_name,
-        source_product_version=curve.product_version,
-        source_id=source_id,
-        content_hash=source_hash,
-        as_of_date=curve.as_of_date,
-        window_start_date=curve.window.start_date,
-        window_end_date=curve.window.end_date,
-        returned_curve_point_count=curve.supportability.returned_curve_point_count,
-        missing_security_ids=curve.supportability.missing_security_ids,
-        curve_points=[
-            AuthoritativeTransactionCostPoint(
-                security_id=point.security_id,
-                transaction_type=point.transaction_type,
-                currency=point.currency,
-                observation_count=point.observation_count,
-                total_notional=point.total_notional,
-                total_cost=point.total_cost,
-                average_cost_bps=point.average_cost_bps,
-                min_cost_bps=point.min_cost_bps,
-                max_cost_bps=point.max_cost_bps,
-                first_observed_date=point.first_observed_date,
-                last_observed_date=point.last_observed_date,
-                sample_transaction_ids=point.sample_transaction_ids[:5],
-            )
-            for point in curve.curve_points[:10]
-        ],
-        reason_codes=[curve.supportability.reason],
-    )
 
 
 def external_treasury_currency_overlay_context(
