+++ tests/unit/dpm/construction/test_source_product_context.py
@@ -17,6 +17,8 @@ from src.api.services.construction_source_product_context import (
     external_order_execution_acknowledgement_context,
     external_treasury_currency_overlay_context,
     source_product_authority_context_updates,
+)
+from src.api.services.construction_transaction_cost_source_context import (
     transaction_cost_context_from_curve,
 )
 from src.core.construction.models import ConstructionAuthorityContext
